refactor equality proofs into separate modules
